+++ src/main/kotlin/com/bredogen/projectenv/core/EnvFileErrorException.kt
@@ -1,7 +1,0 @@
-package com.bredogen.projectenv.core
-
-class EnvFileErrorException : Exception {
-    constructor(message: String?) : super(message)
-    constructor(cause: Throwable?) : super(cause)
-    constructor(message: String?, cause: Throwable?) : super(message, cause)
-}