+++ src/main/kotlin/com/bredogen/projectenv/extensions/JavaRunConfigurationExtension.kt
@@ -1,0 +1,20 @@
+package com.bredogen.projectenv.extensions
+
+import com.bredogen.projectenv.services.ProjectEnvService
+import com.intellij.execution.RunConfigurationExtension
+import com.intellij.execution.configurations.JavaParameters
+import com.intellij.execution.configurations.RunConfigurationBase
+import com.intellij.execution.configurations.RunnerSettings
+
+class JavaRunConfigurationExtension : RunConfigurationExtension() {
+    override fun isApplicableFor(configuration: RunConfigurationBase<*>): Boolean = true
+
+    override fun isEnabledFor(applicableConfiguration: RunConfigurationBase<*>, runnerSettings: RunnerSettings?): Boolean {
+        return ProjectEnvService.getInstance(applicableConfiguration.project).enableRunConfiguration
+    }
+
+    override fun <T : RunConfigurationBase<*>> updateJavaParameters(configuration: T, params: JavaParameters, runnerSettings: RunnerSettings?) {
+        val envService = ProjectEnvService.getInstance(configuration.project)
+        params.env.putAll(envService.getEnvValues())
+    }
+}