+++ src/main/kotlin/com/bredogen/projectenv/providers/EnvProvider.kt
@@ -1,0 +1,11 @@
+package com.bredogen.projectenv.providers
+import com.intellij.openapi.project.Project
+
+interface EnvProvider {
+    fun handleDoubleClick(project: Project) : Boolean
+    fun getEnvValues() : LinkedHashMap<String, String>
+
+    val isValid : Boolean
+    val isFile : Boolean
+    val name : String
+}