+++ src/main/kotlin/com/bredogen/projectenv/providers/EnvProviderFactory.kt
@@ -1,0 +1,8 @@
+package com.bredogen.projectenv.providers
+
+import com.intellij.openapi.project.Project
+
+interface EnvProviderFactory {
+    fun newInstance(params : Map<String, String>): EnvProvider
+    fun createParams(project: Project) : Map<String, String>
+}