+++ src/main/kotlin/com/bredogen/projectenv/providers/EnvSourceException.kt
@@ -1,0 +1,6 @@
+package com.bredogen.projectenv.providers
+
+class EnvSourceException : Exception {
+    constructor(message: String?) : super(message)
+    constructor(cause: Throwable?) : super(cause)
+}